+++ dependency-injection/src/injectable-options.ts
@@ -1,5 +1,11 @@
 import { Scope } from './types';
 
+/**
+ * Stellt Optionen für die Markierung injizierbarer Dienste dar.
+ */
 export interface InjectableOptions {
+  /**
+   * Das Scope, in dem der Dienst verfügbar gemacht werden soll.
+   */
   scope: Scope;
 }
