+++ dependency-injection/src/types.ts
@@ -1,6 +1,21 @@
 import type { DiProvider } from './di-provider';
 
+/**
+ * Stellt einen Konstruktor mit beliebigen Parametern dar.
+ */
 export type Ctor<T> = (new (...params: any[]) => T) | (new () => T);
-export type ConstructedType<T> = T extends Ctor<infer TConstructed> ? TConstructed : never;
-export type ProviderFunction<T> = ((scopedDi: DiProvider) => T) | (() => T);
+
+/**
+ * Holt den Ergebnistyp eines Aufrufs `new T()`, falls `T` ein Konstruktor ist.
+ */
+export type ConstructedType<T extends Ctor<unknown>> = T extends Ctor<infer TConstructed> ? TConstructed : never;
+
+/**
+ * Eine Funktion, die einen injizierbaren Dienst bereitstellt.
+ */
+export type ProviderFunction<T> = (scopedDi: DiProvider) => T;
+
+/**
+ * Scopes / Gültigkeitsbereiche für injizierbare Dienste.
+ */
 export type Scope = 'singleton' | 'scoped' | 'transient';
