+++ time-span/package.json
@@ -33,5 +33,8 @@
   },
   "publishConfig": {
     "access": "public"
+  },
+  "dependencies": {
+    "uninstall": "^0.0.0"
   }
 }
