+++ time-span/test/time-span-format.test.ts
@@ -1,4 +1,4 @@
-import { TimeSpan, TimeSpanFormat } from '../src';
+import { describe, it } from '@jest/globals';
 
 describe('time-span-format', () => {
   it('should format correctly', () => {});
