+++ utility/src/index.ts
@@ -1,4 +1,5 @@
 export * from './utils';
 export * from './extensions';
 export * from './types';
+export * from './meta-data';
 export * from './other';
