+++ utility/src/meta-data/index.ts
@@ -1,0 +1,2 @@
+export * from './meta-data';
+export * from './meta-data-key';
