+++ utility/src/other/index.ts
@@ -1,2 +1,3 @@
 export * from './cache';
 export * from './throw-helper';
+export * from '../meta-data/meta-data'
