+++ utility/tsconfig.json
@@ -18,6 +18,7 @@
     "target": "ES2021",
     "lib": [
       "ES2021",
+      "DOM"
     ]
   },
   "include": [
