+++ utility/tsconfig.test.json
@@ -9,7 +9,7 @@
     ],
   },
   "include": [
-    "./test/**/*.test.ts"
+    "./test/**/*.test.ts",
   ],
   "exclude": [
     "./node_modules",
